+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.3.0](https://github.com/rizen/pixi-glyphs/compare/v4.2.1...v4.3.0) (2025-10-23)
+
 ### [4.2.1](https://github.com/rizen/pixi-glyphs/compare/v4.2.0...v4.2.1) (2025-10-14)
 
 - Fix: Ensure that when topTrim reduces the ascent of a line, the baseHeight is recalculated from the effective ascent (with topTrim applied) plus descent. This makes subsequent lines move up properly without creating gaps.